+++ DfE.GIAP.All/DfE.GIAP.Core.UnitTests/NewsArticles/Infrastructure/Repositories/CosmosNewsArticleReadRepositoryTests.cs
@@ -1,6 +1,6 @@
 ﻿using Dfe.Data.Common.Infrastructure.Persistence.CosmosDb.Handlers.Query;
 using DfE.GIAP.Core.Common.CrossCutting;
-using DfE.GIAP.Core.NewsArticles.Application.UseCases.GetNewsArticles;
+using DfE.GIAP.Core.NewsArticles.Application.Enums;
 using DfE.GIAP.Core.NewsArticles.Infrastructure.Repositories;
 using DfE.GIAP.Core.UnitTests.NewsArticles.Infrastructure.Repositories.TestDoubles;
 using DfE.GIAP.Core.UnitTests.NewsArticles.UseCases;
