+++ DfE.GIAP.All/DfE.GIAP.Core/NewsArticles/Application/UseCases/GetNewsArticles/GetNewsArticlesRequest.cs
@@ -1,4 +1,6 @@
-﻿namespace DfE.GIAP.Core.NewsArticles.Application.UseCases.GetNewsArticles;
+﻿using DfE.GIAP.Core.NewsArticles.Application.Enums;
+
+namespace DfE.GIAP.Core.NewsArticles.Application.UseCases.GetNewsArticles;
 
 /// <summary>
 /// Represents a request to retrieve news articles based on specified search criteria.
@@ -7,13 +9,3 @@
 /// criteria for narrowing down the results, such as keywords, date ranges, or categories.</remarks>
 /// <param name="newsArticleSearchFilter">The filter specifying the search criteria for retrieving news articles.</param>
 public record GetNewsArticlesRequest(NewsArticleSearchFilter newsArticleSearchFilter);
-
-public enum NewsArticleSearchFilter
-{
-    ArchivedWithPublished,
-    ArchivedWithNotPublished,
-    ArchivedWithPublishedAndNotPublished,
-    NotArchivedWithPublished,
-    NotArchivedWithNotPublished,
-    NotArchivedWithPublishedAndNotPublished
-}
