+++ DfE.GIAP.All/DfE.GIAP.Core/NewsArticles/Application/UseCases/GetNewsArticles/GetNewsArticlesUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using DfE.GIAP.Core.Common.Application;
+using DfE.GIAP.Core.NewsArticles.Application.Enums;
 using DfE.GIAP.Core.NewsArticles.Application.Models;
 using DfE.GIAP.Core.NewsArticles.Application.Repositories;
